+++ services/blockchain-connector/shared/sdk/blocksUtils.js
@@ -20,11 +20,10 @@ const path = require('path');
 const { Logger } = require('lisk-service-framework');
 
 const { getNodeInfo } = require('./endpoints_1');
-const { exists, mkdir } = require('../fsUtils');
+const { exists, mkdir, verifyFileChecksum, deleteFileRecursive, extractTarBall } = require('../utils/fs');
 const { downloadFile } = require('../downloadFile');
 
 const config = require('../../config');
-const { verifyFileChecksum, deleteFileRecursive, extractTarBall } = require('../fsUtils');
 
 const logger = Logger();
