+++ services/blockchain-connector/shared/utils/fs.js
